requirements.txt
altair==4.2.0
appdirs==1.4.4
apsw==3.41.2.0
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
entrypoints==0.4
fonttools==4.38.0
fpdf2==2.6.1
gitdb==4.0.10
GitPython==3.1.29
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
gsheetsdb==0.1.13.1
html5lib==1.1
httplib2==0.22.0
identify==2.5.22
idna==3.4
importlib-metadata==5.2.0
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpath-python==1.0.6
jsonschema==4.17.3
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
MarkupSafe==2.1.1
matplotlib==3.6.2
mccabe==0.7.0
mo-dots==4.22.21108
mo-future==3.147.20327
mo-imports==3.149.20327
mo-kwargs==4.22.21108
mo-logs==4.23.21108
moz-sql-parser==4.40.21126
nodeenv==1.7.0
numpy==1.24.0
oauthlib==3.2.2
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
platformdirs==3.2.0
plotly==5.11.0
pluggy==1.0.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.8.0
pyfakefs==5.2.2
Pygments==2.13.0
pygsheets==2.0.6
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
pytest-integration==0.2.2
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
semver==2.13.0
six==1.16.0
smmap==5.0.0
soupsieve==2.4
streamlit==1.26.0
tabulate==0.8.9
tenacity==8.1.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.7
toolz==0.12.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==1.26.13
validators==0.20.0
wcwidth==0.2.6
webencodings==0.5.1
wrapt==1.15.0
zipp==3.11.0
